+++ packages/node/src/indexer/api.service.ts
@@ -2,14 +2,16 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import { ConnectionInfo } from '@near-js/providers/lib/fetch_json';
-import { Inject, Injectable, OnApplicationShutdown } from '@nestjs/common';
+import { Inject, Injectable } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
+
 import {
   getLogger,
   NodeConfig,
   profilerWrap,
   ConnectionPoolService,
   ApiService as BaseApiService,
+  IBlock,
 } from '@subql/node-core';
 import * as Near from 'near-api-js';
 import {
@@ -27,16 +29,13 @@ import * as NearUtil from '../utils/near';
 import { NearApiConnection } from './nearApi.connection';
 import { BlockContent } from './types';
 
-const GENESIS_BLOCK = 9_820_210;
-const MAX_RECONNECT_ATTEMPTS = 5;
-
 const logger = getLogger('api');
 
 @Injectable()
 export class ApiService extends BaseApiService<
   Near.providers.JsonRpcProvider,
   SafeJsonRpcProvider,
-  BlockContent[]
+  IBlock<BlockContent>[]
 > {
   private fetchBlocksBatches = NearUtil.fetchBlocksBatches;