+++ packages/node/src/indexer/blockDispatcher/near-block-dispatcher.ts
@@ -2,7 +2,8 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import { IBlockDispatcher } from '@subql/node-core';
+import { NearBlock } from '@subql/types-near';
 
-export interface INearBlockDispatcher extends IBlockDispatcher {
+export interface INearBlockDispatcher extends IBlockDispatcher<NearBlock> {
   init(onDynamicDsCreated: (height: number) => Promise<void>): Promise<void>;
 }