+++ packages/node/src/utils/near.spec.ts
@@ -1,14 +1,10 @@
 // Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
+import { IBlock } from '@subql/node-core';
 import * as Near from 'near-api-js';
 import { BlockContent } from '../indexer/types';
-import {
-  filterActions,
-  fetchBlocksBatches,
-  filterReceipt,
-  filterReceipts,
-} from './near';
+import { filterActions, fetchBlocksBatches, filterReceipts } from './near';
 
 jest.setTimeout(30000);
 
@@ -22,14 +18,14 @@ describe('Near api', () => {
   });
 
   describe('Receipt Filters', () => {
-    let block: BlockContent;
+    let block: IBlock<BlockContent>;
 
     beforeAll(async () => {
       [block] = await fetchBlocksBatches(nearApi, [85686945]);
     });
 
     it('Can filter receipts with sender, receiver and signer', () => {
-      const actions = filterReceipts(block.receipts, {
+      const actions = filterReceipts(block.block.receipts, {
         sender: 'a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48.factory.bridge.near',
         receiver: 'dclv2.ref-labs.near',
         signer:
@@ -40,7 +36,7 @@ describe('Near api', () => {
     });
 
     it('Can filter receipts with sender, receiver and signer - filter does not match', () => {
-      const actions = filterReceipts(block.receipts, {
+      const actions = filterReceipts(block.block.receipts, {
         sender: 'a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48.factory.bridge.near',
         receiver: '1234',
         signer:
@@ -52,9 +48,9 @@ describe('Near api', () => {
   });
 
   describe('Action Filters', () => {
-    let block50838341: BlockContent,
-      block83788766: BlockContent,
-      block85686945: BlockContent;
+    let block50838341: IBlock<BlockContent>,
+      block83788766: IBlock<BlockContent>,
+      block85686945: IBlock<BlockContent>;
     beforeAll(async () => {
       [block50838341, block83788766, block85686945] = await fetchBlocksBatches(
         nearApi,
@@ -63,7 +59,7 @@ describe('Near api', () => {
     });
 
     it('Can filter FunctionCall actions', () => {
-      const actions = filterActions(block50838341.actions, {
+      const actions = filterActions(block50838341.block.actions, {
         type: 'FunctionCall',
         methodName: 'add_oracle',
         receiver: 'priceoracle.near',
@@ -74,7 +70,7 @@ describe('Near api', () => {
 
     // TODO find transaction of this type
     it.skip('Can filter Stake actions', () => {
-      const actions = filterActions(block50838341.actions, {
+      const actions = filterActions(block50838341.block.actions, {
         type: 'Stake',
         publicKey: '',
       });
@@ -83,7 +79,7 @@ describe('Near api', () => {
     });
 
     it('Can filter AddKey actions', () => {
-      const actions = filterActions(block83788766.actions, {
+      const actions = filterActions(block83788766.block.actions, {
         type: 'AddKey',
         publicKey: 'ed25519:7uCZ9oSNZgLWApTgnYyAMTbqtnsmBYQ2Zn2p5XHnibJ2',
       });
@@ -92,7 +88,7 @@ describe('Near api', () => {
     });
 
     it('Can filter DeleteKey actions', () => {
-      const actions = filterActions(block83788766.actions, {
+      const actions = filterActions(block83788766.block.actions, {
         type: 'DeleteKey',
         publicKey: 'ed25519:9KNsmHHBPfoHZZHKjmBzRQsRDEe9EyNeptamK21hGK9a',
       });
@@ -101,7 +97,7 @@ describe('Near api', () => {
     });
 
     it('Can filter actions with receipt filters', () => {
-      const actions = filterActions(block85686945.actions, {
+      const actions = filterActions(block85686945.block.actions, {
         type: 'FunctionCall',
         methodName: 'ft_on_transfer',
         receiver: 'dclv2.ref-labs.near',
@@ -114,7 +110,7 @@ describe('Near api', () => {
     it.skip('Can filter DeleteAccount actions', async () => {
       const [block] = await fetchBlocksBatches(nearApi, [50838341]);
 
-      const actions = filterActions(block.actions, {
+      const actions = filterActions(block.block.actions, {
         type: 'DeleteAccount',
         beneficiaryId: '',
       });
@@ -124,7 +120,7 @@ describe('Near api', () => {
 
     it('Can filter SignedDelegate actions', async () => {
       const [delegateBlock] = await fetchBlocksBatches(nearApi, [100051916]);
-      const actions = filterActions(delegateBlock.actions, {
+      const actions = filterActions(delegateBlock.block.actions, {
         type: 'Delegate',
       });
       expect(actions.length).toBe(1);