+++ packages/types/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Update `@subql/types-core`
+
+
 ## [3.1.3] - 2024-02-07
 ### Changed
 - Update `@subql/types-core`